Refactor API
