+++ priv/react_app/src/API/users.js
@@ -1,15 +1,8 @@
-import { getJwt } from '../Util/JwtHelper';
-import { handleHttpResponse } from './index';
+import { authFetch, handleHttpResponse } from './utils';
 
 export default {
   /**
    * Gets the current user
    */
-  getMe: () =>
-    fetch('/api/me', {
-      headers: new Headers({
-        'Content-Type': 'application/json',
-        Authorization: `Bearer ${getJwt()}`
-      })
-    }).then(handleHttpResponse)
+  getMe: () => authFetch('/api/me').then(handleHttpResponse)
 };
