+++ web/controllers/company_controller.ex
@@ -7,7 +7,7 @@ defmodule Nexpo.CompanyController do
   plug EnsurePermissions, [handler: Nexpo.SessionController,
                            one_of: [%{default: ["read_all"]},
                                     %{default: ["read_companies"]}]
-                          ] when action in [:index, :show]
+                          ] when action in [:show]
   plug EnsurePermissions, [handler: Nexpo.SessionController,
                            one_of: [%{default: ["write_all"]},
                                     %{default: ["write_companies"]}]
